Create screen package
