+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/presentation/PresentationModule.kt
@@ -1,8 +1,8 @@
 package com.igorwojda.showcase.feature.album.presentation
 
 import coil.ImageLoader
-import com.igorwojda.showcase.feature.album.presentation.albumdetail.AlbumDetailViewModel
-import com.igorwojda.showcase.feature.album.presentation.albumlist.AlbumListViewModel
+import com.igorwojda.showcase.feature.album.presentation.screen.albumdetail.AlbumDetailViewModel
+import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.module
