+++ feature_album/src/test/kotlin/com/igorwojda/showcase/feature/album/presentation/albumdetail/AlbumDetailViewModelTest.kt
@@ -1,6 +1,7 @@
 package com.igorwojda.showcase.feature.album.presentation.albumdetail
 
 import com.igorwojda.showcase.feature.album.domain.usecase.GetAlbumUseCase
+import com.igorwojda.showcase.feature.album.presentation.screen.albumdetail.AlbumDetailViewModel
 import com.igorwojda.showcase.library.testutils.CoroutinesTestDispatcherExtension
 import com.igorwojda.showcase.library.testutils.InstantTaskExecutorExtension
 import io.mockk.mockk
