+++ feature_favourite/src/main/res/navigation/feature_favourite_nav_graph.xml
@@ -6,6 +6,6 @@
         app:startDestination="@id/favouriteFragment">
     <fragment
             android:id="@+id/favouriteFragment"
-            android:name="com.igorwojda.showcase.feature.favourite.presentation.favourite.FavouriteFragment"
+            android:name="com.igorwojda.showcase.feature.favourite.presentation.screen.favourite.FavouriteFragment"
             android:label="FavouriteFragment" />
 </navigation>
