+++ feature_profile/src/main/kotlin/com/igorwojda/showcase/feature/profile/presentation/screen/profile/ProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.igorwojda.showcase.feature.favourite.presentation.favourite
+package com.igorwojda.showcase.feature.profile.presentation.screen.profile
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,12 +10,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.fragment.app.Fragment
 import com.igorwojda.showcase.base.presentation.compose.composable.UnderConstructionAnim
 
-class FavouriteFragment : Fragment() {
+class ProfileFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return ComposeView(requireContext()).apply {
             setContent {
-                FavouriteScreen()
+                ProfileScreen()
             }
         }
     }
@@ -23,6 +23,6 @@ class FavouriteFragment : Fragment() {
 
 @Preview
 @Composable
-private fun FavouriteScreen() {
+private fun ProfileScreen() {
     UnderConstructionAnim()
 }
